+++ de.metas.fresh/de.metas.fresh.base/src/main/sql/postgresql/system/70-de.metas.fresh/5439020_sys_09783_Salesgroups_View_And_Function_Modified.sql
@@ -92,3 +92,57 @@ where i.isSotrx = 'Y'and
 
 
 ;
+
+
+
+DROP FUNCTION IF EXISTS report.Salesgroups (IN DateFrom timestamp without time zone, In DateTo timestamp without time zone);
+
+DROP TABLE IF EXISTS report.Salesgroups_Report_Sub;
+
+
+CREATE TABLE report.Salesgroups_Report_Sub
+(
+	
+	productsalesgroup character varying(60),
+	uom character varying(60),
+	qtyinvoiceddiscounter numeric,			
+	qtyinvoicedgastro numeric,			
+	qtyinvoiceddetailhandel	numeric,	
+	revenuediscounter numeric,			
+	revenuegastro numeric,			
+	revenuedetailhandel numeric
+	
+)
+WITH (
+	OIDS=FALSE
+);
+
+
+CREATE FUNCTION report.Salesgroups(IN DateFrom timestamp without time zone, In DateTo timestamp without time zone) RETURNS SETOF report.Salesgroups_Report_Sub AS
+$BODY$
+SELECT
+
+	productsalesgroup,
+	uom.name,
+	
+	sum(qtyinvoiceddiscounter),		
+	sum(qtyinvoicedgastro),	
+	sum(qtyinvoiceddetailhandel),	
+	sum(revenuediscounter),		
+	sum(revenuegastro),		
+	sum(revenuedetailhandel)
+	
+FROM report.RV_Salesgroups rv
+JOIN C_UOM uom on uom.C_UOM_ID = rv.uom
+WHERE rv.DateInvoiced >= $1 AND rv.DateInvoiced <= $2
+GROUP BY 	
+	productsalesgroup, uom.name
+	
+	
+	
+	
+	
+ORDER BY
+	productsalesgroup ASC$BODY$
+LANGUAGE sql STABLE;
+